requirements.txt
web3>=6.15.0
eth-account>=0.11.0
requests>=2.31.0
pandas>=2.2.0
numpy>=1.26.0
cryptography>=42.0.0
python-dotenv>=1.0.0
py-solc-x>=2.0.0
